# Avoid allocations for properties arrays for common cases