These arrays are not stored further and are not used by user code.