@sungam3r 2022-12-04
Need to fix `Serilog.Tests.MethodOverloadConventionTests.ILoggerDefaultMethodsShouldBeInSyncWithLogger` 😕 to skip new code.