@sungam3r 2022-12-04
I must admit that this is an unusual test. I have never seen such tests before. Now it exist on
```cs
if (ifaceBytes.Length != classBytes.Length)
{
    return false;
}
```
since `ifaceBytes` is 39 bytes long and `classBytes` is 71 bytes long (for Void Write[T](Serilog.Events.LogEventLevel, System.String, T for example). Interfaces cannot contain instance fields. Even if I add static field in `ILogger` methods are still have different length. What should this test look like now?