@nblumhardt 2023-01-24
I think we'd need to be careful about the cost and concurrency impact of the memory barriers inserted by CompareExchange before taking this route