@sungam3r 2023-01-25
`dotnet test -c Release -f net7.0 --filter "FullyQualifiedName=Serilog.PerformanceTests.Harness.Pipeline"`

``` ini

BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.22621.1105)
AMD Ryzen 7 6800H with Radeon Graphics, 1 CPU, 16 logical and 8 physical cores
.NET SDK=7.0.102
  [Host]     : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2


```

before

|       Method |     Mean |   Error |  StdDev |   Gen0 | Allocated |
|------------- |---------:|--------:|--------:|-------:|----------:|
| EmitLogEvent | 155.6 ns | 3.14 ns | 4.99 ns | 0.0448 |     376 B |

after


|       Method |     Mean |   Error |  StdDev |   Gen0 | Allocated |
|------------- |---------:|--------:|--------:|-------:|----------:|
| EmitLogEvent | 152.2 ns | 2.45 ns | 2.17 ns | 0.0410 |     344 B |