@sungam3r 2023-05-21
Rebased on dev. Still need help with `Serilog.Tests.MethodOverloadConventionTests.ILoggerDefaultMethodsShouldBeInSyncWithLogger`