@sungam3r 2023-06-27
Rebased on dev. @nblumhardt Could you please help with `Serilog.Tests.MethodOverloadConventionTests.ILoggerDefaultMethodsShouldBeInSyncWithLogger`?