@epeshk 2023-08-22
The approach of using the single `_reusableArrayOfX` field for caching may cause contention between threads, as atomic writes [do not scale well](https://www.1024cores.net/home/lock-free-algorithms/first-things-first). And `Interlocked.Exchange` is write access.

The example of contention-free cache implementation can be found in [dotnet/runtime](https://github.com/dotnet/runtime/blob/4a5695b6ea0e461e6527de057420872369345282/src/libraries/System.Private.CoreLib/src/System/Runtime/CompilerServices/PoolingAsyncValueTaskMethodBuilderT.cs#L302) `PoolingAsyncValueTaskMethodBuilder<T>` . This is a two-level cache, with per-thread and per-core slots.

For runtimes with `Span<T>` support, memory overhead can be reduced by allocating on the stack, instead of using an array on the heap.