@sungam3r 2024-03-17
Hi. I was away from actively supporting OSS for a long time. I like supposed solution, did not know about such API. @epeshk I have one question. I look at your code and it seems to me that boxing should occur on such lines for value types.
![изображение](https://github.com/serilog/serilog/assets/21261007/48c1c1b4-b845-465b-ba9f-9198329e3ede)

I understand that the whole purpose of out PRs is to eliminate temporary **array** allocations for logged items of that array. Boxing of each item is out of scope of our intentions. So in both cases boxing takes place, right?